Visual improvement of Envelope selector panel on GUI.
Numbers have been added to the Envelope selector to clarify to the user how the active Envelope can be changed. The mouse cursor now changes when dragging from the panel. drawTextBox has been updated to be more customizable and takes an object of new type TextBoxInfo instead of a list of arguments.