Fix four minor bugs
FrequencyPanels do now calculate the phase for the correct LFO instead of always for the active one.

The last point in ShapeEditors can no longer be deleted.

The active state of links is now stored outside IControls and is therefore not lost after the UI is closed.

The indices of active links are now recovered in UnserializeState.