Turn ShapePoint linked list to vector
The linked list of ShapePoints was turned into a vector. All
methods using pointers to ShapePoints do now operate with indices.
It was necessary to move the full declaration of ShapePoint and
ModulatedParameter to the header file and reveal these classes to
other modules.

Removed unused parameters and methods from
src/UDShaperElements/ShapeEditor.cpp.

Tested the performance with the new implementation.