Add binary search to ShapeEditor::forward
ShapeEditor::forward now uses a combination of binary search and
the previously used iterative approach to find moduated x-positions
of points. This is in most cases significantly faster and scales
better.

Measured processing time and updated performance.md.

In some places of ShapeEditor, I replaced code to insert points
with the method insertPointAt, which does now return the index at
which the point has been inserted.