Fix two bugs
First bug skipped points when unserializing ShapeEditor.
Second bug made it possible to connect the same LFO multiple times
to a single ModulatedParameter.

The bug that caused strange numbers to be added to the plugin name
in FL Studio was on host side. After clearing some preset files
from the FL Studio folder, it disappeared.