Improve graph editing user experience
A smoothing function has been applied to the user input when
editing the interpolation shape between two points on ShapeEditor.
The mouse cursor is now hidden while dragging and jumps back to
the correct position.

There is no maximum for the power of interpolation anymore, but
small bounds in the ModulatedParameter curveCenterPosY, making sure
no infinite power is created. Fixed a bug in
ModulatedParameter::get which bypassed the clamping to [min, max].

Added minor comments to ShapeEditor::forward for more clarity.