Enable dynamic adding and hiding of LFO curves
A new LFO is shown when the last displayed LFO is connected to a
parameter. If the last two LFOs are not connected to any
parameter, the last one is hidden. This holds unless the number
of LFOs would be outside the range
[MIN_NUMBER_LFOS, MAX_NUMBER_LFOS] defined in src/config.h.

Removed the unused member numberLFOs from LFOController.
LFOSelectorControl::numberLFOs is used instead to keep track of
the number of displayed LFOs.

At a certain number of LFOs, the selector panel becomes too crowded
and elements start overlapping. This will be adressed when the UI
gets a general makeover.