Fix minor bugs and improve modulation popup menus
- Bug 1: Modulating the x-position of the last ShapePoint of a
ShapeEditor crashed the plugin. Fixed by disabling the option to
modulate this parameter. Menu option to modulate parameters that
are already linked has been disabled too. Also updated point
highlighting to not consider the x-parameter of the last point.

- Bug 2: Hovering over knobs highlighted the points connected to
the first LFO, not the current one. Removed the member kIdx from
LinkKnobInputLayer and added modIdx, which is the index of the
modulation link obtained from the visual layer when needed.

- Bug 3: ModulatedParameters were always updating LFOIsConnected
in removeModulator instead of only on success, resulting in
incorrect internal state.

- Bug 4: Deleting points not connected to any Modulation links
crashed the plugin.

- Bug 5: Disconnecting modulation links did not reset the internal
amplitudes correctly.