Improve Up/Down processing
I learned that hosts do not parallelize individual plugin
instances, so the amplitude of the last sample of a block is now
stored in UDShaper and accessible to the next block.

Added an internal member for the distortion mode, which is
updated on parameter changes. The unused member TopMenuBar::mode
has been removed.