Added piecewise normalized distortion
A button to toggle the normalization was added to TopMenuBar.
While active, input audio will be normalized to the next minima,
maxima or zero points. This does currently result in audio bugs at
block intersections and requires a lookahead buffer and latency to
work without bugs. The buffer will be implemented in a separate
commit.

Updated UDShaper::ProcessBlock to have one loop for all distortion
modes. Normalization happens in this loop before processing.

Changed the type of some variables from float to iplug::sample,
which can be either double or float depending on the configuration.