Add audio buffer to properly normalize input audio
If normalization is active, the audio is written to a buffer and
processed once 674 samples are loaded to the buffer. Normalized
distortion works as intended without audio bugs, but has high CPU
usage. Next step is to find improvements to findPOI.

modulationStep is now only called if the host is playing.